+++ packages/ccd-js-gen/bin/ccd-js-gen.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const cli = require('../dist/src/cli.js');
+const cli = require('../lib/src/cli.js');
 cli.main();